requirements.txt
discord.py>=2.3.0
google-genai>=0.2.0
python-dotenv>=1.0.0
langdetect>=1.0.9
psutil>=5.9.0
aiohttp>=3.8.0